vmling.o builds

+++ core/bignum/mpi.c
@@ -3098,7 +3098,7 @@ int      s_mp_tovalue(char ch, int r)
   if(r > 36)
     xch = ch;
   else
-    xch = toupper(ch);
+    xch = toupper((int)ch);
 
   if(isdigit(xch))
     val = xch - '0';
@@ -3143,7 +3143,7 @@ char     s_mp_todigit(int val, int r, int low)
   ch = s_dmap_1[val];
 
   if(r <= 36 && low)
-    ch = tolower(ch);
+    ch = tolower((int)ch);
 
   return ch;
